+++ argutil/argutil.py
@@ -18,8 +18,7 @@ from sys import exit
 from .deepcopy import deepcopy
 from .primitives import primitives
 import logging
-
-VERSION = '1.1.4'
+import jsonschema
 
 logger = logging.getLogger('argutil')
 logger.setLevel(logging.ERROR)
@@ -42,6 +41,37 @@ def get_file(**kwargs):
     return inspect.stack()[stackdepth].filename
 
 
+def load(json_file, mode='a'):
+    if mode in 'ar':
+        if os.path.isfile(json_file):
+            with open(json_file, 'r') as f:
+                return json.load(f)
+        elif mode == 'r':
+            raise FileNotFoundError('file could not be read: ' + json_file)
+    if mode in 'wca':
+        return {}
+    raise ValueError('Unknown file mode "{}"'.format(mode))
+
+
+def save(json_data, json_file):
+    with open(json_file, 'w') as f:
+        f.write(json.dumps(json_data, indent=2))
+
+
+def validate(json_data_or_file):
+    if isinstance(json_data_or_file, dict):
+        json_data = json_data_or_file
+    elif (
+        isinstance(json_data_or_file, str) and
+        os.path.isfile(json_data_or_file)
+    ):
+        json_data = load(json_data_or_file, 'r')
+    jsonschema.validate(json_data, commandline_schema)
+    return json_data
+
+
+with WorkingDirectory(__file__):
+    commandline_schema = load('commandline.schema')
 GLOBAL_ENV = {}
 
 
@@ -73,13 +103,12 @@ class ParserDefinition(object):
             argutil_dir = os.path.dirname(argutil_path)
             template_path = os.path.join(argutil_dir, defaults.TEMPLATE_FILE)
             shutil.copy2(template_path, filepath)
+        with WorkingDirectory(filepath):
+            definitions_file = os.path.abspath(definitions_file)
+            defaults_file = os.path.abspath(defaults_file)
         if not os.path.isfile(definitions_file):
             save({'modules': {}}, definitions_file)
-        json_data = load(definitions_file)
-        if 'modules' not in json_data:
-            raise KeyError(
-                '{} does not contain key "modules"'.format(definitions_file)
-            )
+        json_data = validate(definitions_file)
         if module in json_data['modules']:
             if fail_if_exists:
                 raise KeyError('module already defined')
@@ -104,8 +133,9 @@ class ParserDefinition(object):
             filepath = os.path.abspath(filepath)
         self.filepath = filepath
         self.module = get_module(filepath)
-        self.definitions_file = definitions_file
-        self.defaults_file = defaults_file
+        with WorkingDirectory(filepath):
+            self.definitions_file = os.path.abspath(definitions_file)
+            self.defaults_file = os.path.abspath(defaults_file)
         self.env = env or {}
 
     def callable(self, name=None):
@@ -114,30 +144,27 @@ class ParserDefinition(object):
             return function
         return decorator
 
-    def load(self, json_file, mode='a'):
-        with WorkingDirectory(self.filepath):
-            return load(json_file, mode)
-
-    def save(self, json_data, json_file):
-        with WorkingDirectory(self.filepath):
-            return save(json_data, json_file)
+    def delete(self):
+        json_data = load(self.definitions_file)
+        del json_data['modules'][self.module]
+        save(json_data, self.definitions_file)
+        json_data = load(self.defaults_file)
+        if self.module in json_data:
+            del json_data[self.module]
+            save(json_data, self.defaults_file)
 
     def add_example(
         self,
         usage,
         description='',
     ):
-        if not isinstance(usage, str):
-            raise ValueError('usage must be a string!')
-        if not isinstance(description, str):
-            raise ValueError('description must be a string!')
-        json_data = self.load(self.definitions_file)
+        json_data = load(self.definitions_file)
         example = {
             'usage': usage,
             'description': description
         }
         json_data['modules'][self.module]['examples'].append(example)
-        self.save(json_data, self.definitions_file)
+        save(validate(json_data), self.definitions_file)
 
     def add_argument(
         self,
@@ -145,7 +172,7 @@ class ParserDefinition(object):
         short=None,
         **kwargs
     ):
-        json_data = self.load(self.definitions_file)
+        json_data = load(self.definitions_file)
         arg = {}
         if short is not None:
             arg['short'] = short
@@ -174,10 +201,10 @@ class ParserDefinition(object):
         arg['help'] = help
 
         json_data['modules'][self.module]['args'].append(arg)
-        self.save(json_data, self.definitions_file)
+        save(validate(json_data), self.definitions_file)
 
     def set_defaults(self, **kwargs):
-        json_data = self.load(self.defaults_file)
+        json_data = load(self.defaults_file)
         if self.module not in json_data:
             json_data[self.module] = {}
         module = json_data[self.module]
@@ -191,25 +218,25 @@ class ParserDefinition(object):
                 m = m[k_parent]
                 k = k[index + 1:]
             m[k] = v
-        self.save(json_data, self.defaults_file)
+        save(json_data, self.defaults_file)
 
     def get_defaults(self):
-        json_data = self.load(self.defaults_file)
+        json_data = load(self.defaults_file)
         return json_data.get(self.module, {})
 
     def config(self, configs=None):
         if configs:
-            defaults = {}
+            module_defaults = {}
             for k, v in [__split_any__(kv, '=:') for kv in configs]:
                 if v[0] == '[' and v[-1] == ']':
                     v = [__parse_value__(s) for s in v[1:-1].split(',')]
                 else:
                     v = __parse_value__(v)
-                defaults[k] = v
-            self.set_defaults(**defaults)
+                module_defaults[k] = v
+            self.set_defaults(**module_defaults)
         else:
-            defaults = self.get_defaults()
-            items = list(defaults.items())
+            module_defaults = self.get_defaults()
+            items = list(module_defaults.items())
             configs = []
             while items:
                 k, v = items.pop(0)
@@ -222,77 +249,49 @@ class ParserDefinition(object):
             return configs
 
     def get_parser(self, env=None):
-        with WorkingDirectory(self.filepath):
-            if not os.path.isfile(self.definitions_file):
-                logger.error(
-                    'Argument definition file "{}" not found!'.format(
-                        self.definitions_file
-                    )
+        if not os.path.isfile(self.definitions_file):
+            logger.error(
+                'Argument definition file "{}" not found!'.format(
+                    self.definitions_file
                 )
-                exit(1)
-            if env is None:
-                env = {}
-
-            json_data = load(self.definitions_file)
-
-            if 'modules' not in json_data:
-                return ArgumentParser(
-                    epilog='{} does not contain any modules'.format(
-                        self.definitions_file
-                    )
-                )
-            json_data = json_data['modules']
-            if self.module not in json_data:
-                return ArgumentParser(
-                    epilog='No entry for {} in {}'.format(
-                        self.module,
-                        self.definitions_file
-                    )
-                )
-            json_data = json_data[self.module]
-
-            if os.path.isfile(self.defaults_file):
-                defaults = load(self.defaults_file)
-                if self.module in defaults:
-                    defaults = defaults[self.module]
-                else:
-                    defaults = {}
-            else:
-                defaults = {}
-            env = dict(env)
-            for k, v in GLOBAL_ENV.items():
-                env[k] = v
-            for k, v in self.env.items():
-                env[k] = v
-
-            return __build_parser__(
-                self.module,
-                json_data,
-                defaults=defaults,
-                env=env
             )
+            exit(1)
+        if env is None:
+            env = {}
+
+        json_data = validate(self.definitions_file)['modules']
+        if self.module not in json_data:
+            raise KeyError(
+                'No entry for {} in {}'.format(
+                    self.module,
+                    self.definitions_file
+                )
+            )
+        json_data = json_data[self.module]
+
+        if os.path.isfile(self.defaults_file):
+            module_defaults = load(self.defaults_file)
+            if self.module in module_defaults:
+                module_defaults = module_defaults[self.module]
+            else:
+                module_defaults = {}
+        else:
+            module_defaults = {}
+        env = dict(env)
+        for k, v in GLOBAL_ENV.items():
+            env[k] = v
+        for k, v in self.env.items():
+            env[k] = v
+
+        return __build_parser__(
+            self.module,
+            json_data,
+            module_defaults=module_defaults,
+            env=env
+        )
 
 
-def load(json_file, mode='a'):
-    if mode in 'ar':
-        if os.path.isfile(json_file):
-            with open(json_file, 'r') as f:
-                return json.load(f)
-        elif mode == 'r':
-            raise FileNotFoundError('file could not be read: ' + json_file)
-    if mode in 'wca':
-        return {}
-    raise ValueError('Unknown file mode "{}"'.format(mode))
-
-
-def save(json_data, json_file):
-    with open(json_file, 'w') as f:
-        f.write(json.dumps(json_data, indent=2))
-
-
-def __split_any__(text, delimiters=None):
-    if delimiters is None:
-        delimiters = []
+def __split_any__(text, delimiters):
     parts = [text]
     for delim in delimiters:
         parts = [p for ps in parts for p in ps.split(delim)]
@@ -319,9 +318,7 @@ def __parse_value__(value):
             return v
 
 
-def __add_argument_to_parser__(parser, param, env=None):
-    if env is None:
-        env = {}
+def __add_argument_to_parser__(parser, param, env):
     param = dict(param)
     if 'help' in param:
         if param['help'] is None:
@@ -339,8 +336,6 @@ def __add_argument_to_parser__(parser, param, env=None):
                 param['type'] = primitives[func]
             except KeyError:
                 param['type'] = globals()[func]
-    if 'long' not in param:
-        raise ValueError('args must contain "long" key')
     long_form = param['long']
     del param['long']
     if 'short' in param:
@@ -354,15 +349,11 @@ def __add_argument_to_parser__(parser, param, env=None):
 def __add_example_to_parser__(parserArgs, example):
     if 'epilog' not in parserArgs:
         parserArgs['epilog'] = 'examples:'
-    parserArgs['epilog'] += '\n    {:<44}{}'.format(*(example.values()))
+    parserArgs['epilog'] += '\n    {usage:<44}{description}'.format(**example)
 
 
-def __build_parser__(name, definition, defaults=None, env=None,
+def __build_parser__(name, definition, module_defaults, env,
                      subparsers=None, templates=None, parents=None):
-    if defaults is None:
-        defaults = {}
-    if env is None:
-        env = {}
     if templates is None:
         templates = {}
     if parents is None:
@@ -373,7 +364,7 @@ def __build_parser__(name, definition, defaults=None, env=None,
     if 'template' in definition:
         template_name = definition['template']
         if template_name not in templates:
-            raise ValueError('unknown template ' + template_name)
+            raise KeyError('unknown template ' + template_name)
         template = templates[template_name]
         if 'examples' in template:
             for example in template['examples']:
@@ -388,7 +379,8 @@ def __build_parser__(name, definition, defaults=None, env=None,
     if subparsers is None:
         parser = ArgumentParser(**parserArgs)
     else:
-        for k in ['help', 'aliases']:
+        valid_kwargs = ['help', 'aliases']
+        for k in valid_kwargs:
             if k in definition:
                 parserArgs[k] = definition[k]
         parser = subparsers.add_parser(name, **parserArgs)
@@ -396,7 +388,8 @@ def __build_parser__(name, definition, defaults=None, env=None,
             parser.set_defaults(func=env[name])
         else:
             def usage(*args, **kwargs):
-                parser.parse_args(parents + [name, '-h'])
+                parser.print_help()
+                return 0
             parser.set_defaults(func=usage)
     if 'args' in definition:
         for param in definition['args']:
@@ -408,21 +401,21 @@ def __build_parser__(name, definition, defaults=None, env=None,
                 __add_argument_to_parser__(parser, param, env)
 
     # Apply default values
-    for k, v in defaults.items():
+    for k, v in module_defaults.items():
         try:
-            defaults[k] = v.format(**env)
+            module_defaults[k] = v.format(**env)
         except AttributeError:
             continue
-    parser.set_defaults(**defaults)
+    parser.set_defaults(**module_defaults)
 
     if 'templates' in definition:
         templates = dict(templates)
         for k, v in definition['templates'].items():
             if 'parent' in v:
                 parent_name = v['parent']
-                if parent_name not in templates:
-                    raise ValueError('unknown parent template ' + parent_name)
-                new_v = deepcopy(templates[parent_name])
+                if parent_name not in definition['templates']:
+                    raise KeyError('unknown parent template ' + parent_name)
+                new_v = deepcopy(definition['templates'][parent_name])
                 for k2, v2 in v.items():
                     if k2 not in new_v:
                         new_v[k2] = []
@@ -443,8 +436,8 @@ def __build_parser__(name, definition, defaults=None, env=None,
     if 'modules' in definition:
         subparsers = parser.add_subparsers(dest='command')
         for submodule_name, submodule in definition['modules'].items():
-            if submodule_name in defaults:
-                sub_defaults = defaults[submodule_name]
+            if submodule_name in module_defaults:
+                sub_defaults = module_defaults[submodule_name]
             else:
                 sub_defaults = {}
             __build_parser__(