+++ tests/get_parser_test.py
@@ -2,6 +2,7 @@ import unittest
 from .helper import WD, tempdir
 import argutil
 from argutil import ParserDefinition
+from jsonschema import ValidationError
 
 DEFINITIONS_FILE = argutil.defaults.DEFINITIONS_FILE
 DEFAULTS_FILE = argutil.defaults.DEFAULTS_FILE
@@ -30,10 +31,10 @@ class GetParserTest(unittest.TestCase):
             'l_str': lambda s: str(s).lower(),
         }
         with argutil.WorkingDirectory(WD):
-            cls.parser_def = ParserDefinition.create(cls.filepath)
+            parser_def = ParserDefinition.create(cls.filepath)
 
             def add_arg(name, **kwargs):
-                return cls.parser_def.add_argument(name, **kwargs)
+                return parser_def.add_argument(name, **kwargs)
             add_arg('positional')
             add_arg('positional_list', nargs='*', type=int)
             add_arg('--flag', action='store_true', help='boolean flag option')
@@ -45,10 +46,16 @@ class GetParserTest(unittest.TestCase):
                 default=[]
             )
             add_arg('--env-type', type='l_str')
-            cls.parser_def.add_example(
+            parser_def.env['hex'] = lambda s: hex(int(s))
+            add_arg('--def-env-type', type='hex')
+            argutil.GLOBAL_ENV['u_str'] = lambda s: s.upper()
+            add_arg('--global-env-type', type='u_str')
+            add_arg('--default-from-file')
+            parser_def.add_example(
                 'test --flag', 'positional w/ optional flag'
             )
-            cls.parser = cls.parser_def.get_parser(cls.env)
+            parser_def.set_defaults(default_from_file=123)
+            cls.parser = parser_def.get_parser(env=cls.env)
 
     def get_opts(self, *args):
         with argutil.WorkingDirectory(WD):
@@ -107,6 +114,58 @@ class GetParserTest(unittest.TestCase):
             'abc'
         )
 
+    def test_parser_definition_type_option(self):
+        self.assertIs(
+            self.get_opts('test').def_env_type,
+            None
+        )
+        self.assertEqual(
+            self.get_opts(
+                'test', '--def-env-type', '255'
+            ).def_env_type,
+            '0xff'
+        )
+
+    def test_global_env_type_option(self):
+        self.assertIs(
+            self.get_opts('test').global_env_type,
+            None
+        )
+        self.assertEqual(
+            self.get_opts(
+                'test', '--global-env-type', 'abc'
+            ).global_env_type,
+            'ABC'
+        )
+
+    def test_default_from_file(self):
+        self.assertEqual(self.get_opts('test').default_from_file, 123)
+
+    @tempdir(WD)
+    def test_module_not_in_defaults_file(self):
+        this_def = argutil.ParserDefinition.create('this_script')
+        this_def.add_argument('--foo')
+        this_def.set_defaults(foo='bar')
+        that_def = argutil.ParserDefinition.create('that_script')
+        that_def.add_argument('--foo')
+        that_parser = that_def.get_parser()
+        that_opts = that_parser.parse_args([])
+        self.assertIs(that_opts.foo, None)
+
+    @tempdir(WD)
+    def test_error_bad_definitions_file(self):
+        with open(DEFINITIONS_FILE, 'w') as f:
+            f.write('{"bad_def":{}}')
+        with self.assertRaises(ValidationError):
+            argutil.get_parser('bad_def.py')
+
+    @tempdir(WD)
+    def test_error_non_existent_module(self):
+        with open(DEFINITIONS_FILE, 'w') as f:
+            f.write('{"modules":{}}')
+        with self.assertRaises(KeyError):
+            argutil.get_parser('bad_def.py')
+
     @tempdir(WD)
     def test_error_unknown_type(self):
         filename = 'bad_module.py'
@@ -132,7 +191,7 @@ class GetParserTest(unittest.TestCase):
             }
         }
         argutil.save(json_data, DEFINITIONS_FILE)
-        with self.assertRaises(ValueError):
+        with self.assertRaises(ValidationError):
             parser_def.get_parser()